WIP ZoomableContainer
