+++ src/base/src/AXOpen.VisualComposer/VisualComposerContainer.razor
@@ -1,20 +1,19 @@
 ﻿@using AXSharp.Connector;
 @using Microsoft.AspNetCore.Components.Forms;
 
-@* <div id="zoomAndMoveContainer" style="overflow: hidden; position: relative; width: 100%; height: 50%;">
-    <div style="width: 200px; height: 200px; background-color: lime;"></div>
-    <div style="width: 200px; height: 200px; background-color: red;"></div>
-</div> *@
-
 <AuthorizeView Roles="Administrator">
     <Authorized>
-        <div id="zoomAndMoveContainer" style="position: relative; width: 100%; height: 100%;" ondragover="event.preventDefault();">
-            @if(_imgId != null)
-            {
-                <img id="@_imgId" src="@ImgSrc" style="width: 100%; height: 100%; object-fit: cover;" />
-            }
+        <div class="d-flex justify-content-end my-2">
+            <div style="z-index: 100;">
+                <select class="form-select w-auto" @onchange="(ChangeEventArgs args) => Load(args.Value.ToString())">
+                    @foreach (var item in GetAllFiles())
+                    {
+                        <option value="@item" selected="@(CurrentTemplate is null || CurrentTemplate == "" ? item == "Default" ? "selected" : null : item == CurrentTemplate ? "selected" : null)">@item</option>
+                    }
+                </select>
+            </div>
 
-            <div class="dropdown position-absolute" style="z-index: 100; left: 97%; top: 0%; translate: -100% -50%;">
+            <div class="dropdown ms-4" style="z-index: 100;">
                 <a class="dropdown-toggle text-decoration-none text-black" style="cursor: pointer;" data-bs-toggle="dropdown" aria-expanded="false">
                     <span class="bi bi-pencil-square fs-4" aria-hidden="true" />
                 </a>
@@ -28,25 +27,24 @@
                     <li><a class="dropdown-item" data-bs-toggle="modal" data-bs-target="#setImageModal-@Id">Set image</a></li>
                 </ul>
             </div>
-
-            <div class="position-absolute" style="z-index: 100; left: 90%; top: 0%; translate: -100% -50%;">
-                <select class="form-select w-auto" @onchange="(ChangeEventArgs args) => Load(args.Value.ToString())">
-                    @foreach (var item in GetAllFiles())
-                    {
-                        <option value="@item" selected="@(CurrentTemplate is null || CurrentTemplate == "" ? item == "Default" ? "selected" : null : item == CurrentTemplate ? "selected" : null)">@item</option>
-                    }
-                </select>
-            </div>
-
-            <CascadingValue Value="@this" Name="Parent" IsFixed="true">
-                <CascadingValue Value="@_imgId" Name="ImgId" IsFixed="true">
-                    @foreach (VisualComposerItem child in _children)
-                    {
-                        <VisualComposerItem Origin="@child" />
-                    }
-                </CascadingValue>
-            </CascadingValue>
         </div>
+        <ZoomableContainer>
+            <div ondragover="event.preventDefault();">
+                @if (_imgId != null)
+                {
+                    <img id="@_imgId" src="@ImgSrc" style="width: 100%; height: 100%; object-fit: cover;" />
+                }
+
+                <CascadingValue Value="@this" Name="Parent" IsFixed="true">
+                    <CascadingValue Value="@_imgId" Name="ImgId" IsFixed="true">
+                        @foreach (VisualComposerItem child in _children)
+                        {
+                            <VisualComposerItem Origin="@child" />
+                        }
+                    </CascadingValue>
+                </CascadingValue>
+            </div>
+        </ZoomableContainer>
 
         <div class="modal fade" id="showAllModal-@Id" tabindex="-1" aria-hidden="true">
             <div class="modal-dialog modal-xl modal-dialog-centered modal-dialog-scrollable">
@@ -203,18 +201,18 @@
         </div>
     </Authorized>
     <NotAuthorized>
-        <div style="position: relative; width: 100%; height: 100%;">
-            <img id="@_imgId" src="@ImgSrc" style="width: 100%; height: 100%; object-fit: cover;" />
-            <CascadingValue Value="@this" Name="Parent" IsFixed="true">
-                <CascadingValue Value="@_imgId" Name="ImgId" IsFixed="true">
-                    @foreach (VisualComposerItem child in _children)
-                    {
-                        <VisualComposerItem Origin="@child" />
-                    }
+        <ZoomableContainer>
+            <div style="position: relative; width: 100%; height: 100%;">
+                <img id="@_imgId" src="@ImgSrc" style="width: 100%; height: 100%; object-fit: cover;" />
+                <CascadingValue Value="@this" Name="Parent" IsFixed="true">
+                    <CascadingValue Value="@_imgId" Name="ImgId" IsFixed="true">
+                        @foreach (VisualComposerItem child in _children)
+                        {
+                            <VisualComposerItem Origin="@child" />
+                        }
+                    </CascadingValue>
                 </CascadingValue>
-            </CascadingValue>
-        </div>
+            </div>
+        </ZoomableContainer>
     </NotAuthorized>
 </AuthorizeView>
-
-<script type="text/javascript" src="./_content/AXOpen.VisualComposer/VisualComposerContainer.razor.js"></script>
