+++ src/base/src/AXOpen.VisualComposer/ZoomableContainer.razor
@@ -1,0 +1,7 @@
+﻿<div id="zoomAndMoveContainer" style="overflow: hidden; border: solid 1px black">
+    <div id="zoomAndMoveItem" style="position: relative; transform: scale(1) translate(0px, 0px)">
+        @ChildContent
+    </div>
+</div>
+
+<script type="text/javascript" src="./_content/AXOpen.VisualComposer/ZoomableContainer.razor.js"></script>
